feat: change argument structure to catch more errors
When using the initial API of git-format-staged I found that I ran into
problems that were difficult to immediatly identify due to mistakes in
command line arguments. When I forgot to quote the `--glob` argument
git-format-staged interpreted project files as formatter commands.
I would forget to provide a formatter command, and it was difficult for
git-format-staged to correctly identify and report that problem.

This change swaps file and formatter command arguments so that the
formatter command is now a flag argument that should be quoted, and one
or more files are given as command line arguments. Expanding globs is
now delegated to the shell - git-format-staged expects a list of file
paths correspond with existing files in the repository. In the new API if the
`--formatter` or `-f` flag is missing then git-format-staged can clearly
communicate to the user that the formatter command is missing. If the
user forgets to quote the formatter command then git-format-staged will
see unrecognized flag arguments, or arguments that are given in the file
list position that do not correspond to files on disk. In either case
git-format-stage suggests to the user that the formatter command should
be quoted. If the user quotes a glob instead of letting the shell expand
it then the user gets a similar suggestion. If the user forgets to
provide a file list or a file glob they will see an error clearly
indicating the problem.

BREAKING CHANGE: argument structure has been entirely changed.

Changes resulting from formatting are now applied to working tree files
by default. Use the new `--no-update-working-tree` flag to disable that
behavior. The old `--update-working-tree` flag has been removed.

The formatter command is now given with a flag (`--formatter` or `-f`)
which is required. The command must be quoted if it includes arguments.
The command will be evaluated in a shell, once for each file to be
formatted.

The `--glob` flag has been removed. Files are now given as one or more
positional arguments. The recommendation is to provide one or glob
arguments to be expanded by the shell evaluating the git-format-staged
command. Glob arguments must be quoted - git-format-staged no longer
expands globs internally. Another option is to list specific files on
the command line.