feat: add `--no-write` option which skips applying formatting changes
With the `--no-write` option git-format-staged runs the given formatter
command, but ignores its output and does not make changes to staged
files or to working tree files. This can be used as a dry-run option for
testing, or it can be used to lint staged versions of files without
formatting them. For example:

    git-format-staged --no-write -f 'eslint --stdin >&2' '*.js'