feat: add support for negated file patterns
File patterns can be negated with a '!' character to exclude files from
formatting. File patterns are evaluated from left-to-right: if a file
matches both positive and negative patterns then the right-most pattern
determines whether the file will be formatted.

BREAKING CHANGE: The use of '!' to indicate a negative file pattern
means that files whose names begin with a literal '!' cannot be
referenced literally in the file pattern list. One possible workaround
is to use the single-character wildcard ('?') in place of '!' in file
patterns.