ci: deploy from release branch (#68)
Until now any commits to `master` would automatically trigger a release on npm. This PR changes that behavior so that releases are triggered on merges to a branch called `release`. The configuration also supports maintenance releases for existing major versions by pushing to branches with names of the format `1.x`, `2.x`, etc.

Moves semantic-release configuration to `.releaserc.yaml`.