feat: compare file patterns use relative paths instead of absolute (#124)
Previously git-format-staged normalized file patterns and staged file
paths based on the git repository root to get absolute paths. This
change removes normalization. Staged file paths reported by git are
always relative to the repo root.

Fixes #109 

BREAKING CHANGE: Running git-format-staged in a repo subdirectory with
unquoted file glob patterns, or with literal file paths will not work as
it did before. File patterns must now be relative to the repo root. This
means that it is important to quote shell globs when running in a
subdirectory - otherwise they are expanded relative to the subdirectory
instead of the repo root.